Fix timezone issues in the decagon data processing pipeline #1
Use lubridate package instead of as.Posixct and sfrptime for parsing
characters to dates.

Use defauld UTC timezone.  Timezone is tricky with daylight savings
and some functions pull the local machines timezone when parsing
dates to strings rather than using UTC.  So this seems more consistent.

Fixing the pipeline I discovered a slight error in the tod table
but I'm not sure if it changed anything.